+++ src/confusius/glm/__init__.py
@@ -22,7 +22,6 @@ from confusius.glm._models import (
 )  # exported for user type annotations
 from confusius.glm.first_level import FirstLevelModel
 from confusius.glm.second_level import SecondLevelModel, make_second_level_design_matrix
-from confusius.glm.thresholding import apply_statistical_threshold
 
 __all__ = [
     "FirstLevelModel",
@@ -31,7 +30,6 @@ __all__ = [
     "make_second_level_design_matrix",
     "RegressionResults",
     "Contrast",
-    "apply_statistical_threshold",
     "gamma_difference_hrf",
     "glover_hrf",
     "spm_hrf",